refactor(whl): move/standardize `Taxa::WhatLinksHere`
